requirements.txt
streamlit==1.27.0
torch==2.0.1
torchvision==0.15.2
Pillow>=7.1.0,<10.0.0
matplotlib==3.7.2
numpy==1.24.3 